refactor: typesのgraphを移動
